Adding final changes to ticket
